Allow for a large number of column names in compound indicies
When creating a compound/composite index with 6 columns we noticed that the while loop to create the index name was never ending, causing the php artisan migrate command to hang.

By providing for recognition of large numbers of column in compound indexes we can avoid the while loop completely by writing a sane and unique index name.